Fix Code formatting issues
