+++ tests/Api/Docstore/testdata/petstore.swagger.json
